Multiple change request orders
Palantir/Overwatch - Significant updates to be able to track what the
employee is working on to increase speed of CIFR'ing etc. This is still
only in the ALPHA stage and is only being tested by me.

Collateral Log - Request to have a more accurate collateral log, now a
file is uploaded (this is separate from the program as of this moment)
with the FULL VINS to choose from for the clerk. The clerk then logs
when a certain VIN is received. This will begin BETA testing by the Wolf
Pack